+++ src/main/java/de/uib/configed/gui/swinfopage/PanelSWInfo.java
@@ -421,7 +421,7 @@ public class PanelSWInfo extends JPanel {
 		metaData.put("subject", "report of table");
 		metaData.put("keywords", "software inventory");
 
-		ExporterToPDF pdfExportTable = new ExporterToPDF(panelTable);
+		ExporterToPDF pdfExportTable = new ExporterToPDF(panelTable.getTheTable());
 		pdfExportTable.setClient(hostId);
 		pdfExportTable.setMetaData(metaData);
 		pdfExportTable.setPageSizeA4Landscape();